pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tla-eval"
version = "0.1.0"
description = "A framework for evaluating LLMs on TLA+ specification generation"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "TLA+ Benchmark Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.8"
dependencies = [
    "openai>=1.0.0",
    "anthropic>=0.21.0",
    "pyyaml>=6.0.0",
    "requests>=2.25.0",
    "tqdm>=4.65.0",
    "pandas>=1.5.0",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
    "structlog>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/Qian-Cheng-nju/LLM_Gen_TLA_benchmark_framework"
Repository = "https://github.com/Qian-Cheng-nju/LLM_Gen_TLA_benchmark_framework"
Issues = "https://github.com/Qian-Cheng-nju/LLM_Gen_TLA_benchmark_framework/issues"

[project.scripts]
tla-benchmark = "scripts.run_benchmark:main"
tla-setup = "scripts.setup_tools:main"

[tool.setuptools.packages.find]
include = ["tla_eval*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"